Added ResourceConfiguration types (Currently only DEFAULT is available)

+++ dogan/DoganCell.cpp
@@ -4,10 +4,10 @@
 
 std::string DoganCell::toString(void) {
     std::ostringstream oss;
-    oss << "Type: " << type << "; ContainsRobber: " << containsRobber << ", Number: " << number;
+    oss << "Type: " << static_cast<int>(type) << "; ContainsRobber: " << containsRobber << ", Number: " << number;
     return oss.str();
 
-    // Change when g++ supports C++20
+    // TODO: Change when g++ supports C++20
     // return std::format("Type: {}; ContainsRobber: {}, Number: {}", type, containsRobber, number);
 };
